+++ modules/core/category/src/Hook/ThemeHooks.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Drupal\farm_log_category\Hook;
+namespace Drupal\farm_category\Hook;
 
 use Drupal\Core\Hook\Attribute\Hook;
 
 /**
- * Theme hook implementations for farm_log_category.
+ * Theme hook implementations for farm_category.
  */
 class ThemeHooks {