+++ modules/core/import/modules/csv/tests/src/Kernel/LogCsvImportTest.php
@@ -23,6 +23,7 @@ class LogCsvImportTest extends CsvImportTestBase {
    */
   protected static $modules = [
     'farm_activity',
+    'farm_category',
     'farm_harvest',
     'farm_id_tag',
     'farm_land',